- consolidate bfield analyzers to single file - convert bfield protocol to ABC - rename modules for improved clarity
