+++ mach_eval/analyzers/electromagnetic/resistance.py
